Patch 1206560: allow site extras in reports

+++ edit_report_handler.php
@@ -110,8 +110,8 @@ if ( empty ( $error ) ) {
   }
 }
 
-if ( empty ( $error ) && ! empty ( $report_id ) &&
-  ( $action == "Delete" || $action == translate ( "Delete" ) ) ) {
+if ( empty ( $error ) && ! empty ( $report_id ) && ! empty ( $action )
+     && ( $action == "Delete" || $action == translate ( "Delete" ) ) ) {
   if ( ! dbi_query ( "DELETE FROM webcal_report_template " .
     "WHERE cal_report_id = $report_id" ) )
     $error = translate("Database error") . ": " . dbi_error ();
